Add files via upload
I'm still working on JavaFX tutorials, but this displays a 10 x 10 grid using GridPane and adding 100 rectangle nodes to it. 
I think that before we actually use it I might have to make a GridPane composed of 100 stack panes that each have 1 rectangle in them, but I'm not sure. I need to look at event handling first to see how we can handle processing changes to the grid.